+++ forge/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to the format [MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH](https://mcforge.readthedocs.io/en/1.15.x/conventions/versioning/).
 
 ## [1.17.1-0.32] - 2021.12.23
+### Added
+- Re-added FastFurnace and FastWorkbench compatibility
 ### Fixed
 - Fixed dedicated server lag [#117](https://github.com/TheIllusiveC4/Polymorph/issues/117)
 - Fixed freezes with Applied Energistics 2 pattern terminal [#114](https://github.com/TheIllusiveC4/Polymorph/issues/114)